# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ApplyLeaveRequest:
    properties:
      end_date:
        type: string
      leave_type:
        type: string
      reason:
        type: string
      start_date:
        type: string
    required:
    - end_date
    - leave_type
    - reason
    - start_date
    type: object
  dto.ApproveLeaveRequest:
    properties:
      approved:
        type: boolean
      remarks:
        type: string
    required:
    - approved
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@attendance-workflow.com
    name: API Support
  description: API for managing attendance, leaves, and notifications in an educational
    institution
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Attendance Workflow API
  version: "1.0"
paths:
...
  /leaves:
    get:
      consumes:
      - application/json
      description: Get all leave requests (admin only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              limit:
                type: integer
              page:
                type: integer
              total:
                format: int64
                type: integer
              total_pages:
                format: int64
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all leaves
      tags:
      - leaves
  /leaves/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a leave request (admin only)
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete leave
      tags:
      - leaves
  /leaves/{id}/approve:
    put:
      consumes:
      - application/json
      description: Approve or reject a leave request (admin/faculty/warden only)
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: integer
      - description: Approval decision
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ApproveLeaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve or reject leave
      tags:
      - leaves
  /leaves/apply:
    post:
      consumes:
      - application/json
      description: Submit a leave request
      parameters:
      - description: Leave request details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ApplyLeaveRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                type: object
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Apply for leave
      tags:
      - leaves
  /leaves/my:
    get:
      consumes:
      - application/json
      description: Get leave requests for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              limit:
                type: integer
              page:
                type: integer
              total:
                format: int64
                type: integer
              total_pages:
                format: int64
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my leaves
      tags:
      - leaves
  /leaves/pending:
    get:
      consumes:
      - application/json
      description: Get all pending leave requests (admin/faculty/warden only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              limit:
                type: integer
              page:
                type: integer
              total:
                format: int64
                type: integer
              total_pages:
                format: int64
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get pending leaves
      tags:
      - leaves
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"